ci: forge fmt over all sol files in testdata (#2831)
* run forge fmt over testdata

* add forge fmt check to ci

* run formatter over generated solidity

* use own fmt build to check testdata

* add and use fmt helper

* should only be used with stdin tty

* fmt: fix atty comparison

+++ testdata/cheats/Snapshots.t.sol
@@ -5,8 +5,8 @@ import "ds-test/test.sol";
 import "./Cheats.sol";
 
 struct Storage {
-    uint slot0;
-    uint slot1;
+    uint256 slot0;
+    uint256 slot1;
 }
 
 contract SnapshotTest is DSTest {
@@ -33,7 +33,7 @@ contract SnapshotTest is DSTest {
     }
 
     // tests that snapshots can also revert changes to `block`
-    function testBlockValues() public  {
+    function testBlockValues() public {
         uint256 num = block.number;
         uint256 time = block.timestamp;
         uint256 difficulty = block.difficulty;
@@ -55,5 +55,4 @@ contract SnapshotTest is DSTest {
         assertEq(block.timestamp, time, "snapshot revert for block.timestamp unsuccessful");
         assertEq(block.difficulty, difficulty, "snapshot revert for block.difficulty unsuccessful");
     }
-
 }
